* Make admin layout app layout
