+++ frontend/src/app/admin/admin.module.ts
@@ -1,16 +1,15 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { OverviewComponent } from './overview/overview.component';
-import { PlansComponent } from './plans/plans.component';
-import { OrdersComponent } from './orders/orders.component';
+import {NgModule} from '@angular/core';
+import {CommonModule} from '@angular/common';
+import {OverviewComponent} from './overview/overview.component';
+import {PlansComponent} from './plans/plans.component';
+import {OrdersComponent} from './orders/orders.component';
 import {adminRoutes} from './admin.routes';
 import {UiModule} from '../ui/ui.module';
-import { LayoutComponent } from './layout/layout.component';
-import { UsersComponent } from './users/users.component';
+import {UsersComponent} from './users/users.component';
 import {FormsModule} from '@angular/forms';
 
 @NgModule({
-  declarations: [OverviewComponent, PlansComponent, OrdersComponent, LayoutComponent, UsersComponent],
+  declarations: [OverviewComponent, PlansComponent, OrdersComponent, UsersComponent],
   imports: [
     adminRoutes,
     CommonModule,
@@ -18,4 +17,5 @@ import {FormsModule} from '@angular/forms';
     UiModule
   ]
 })
-export class AdminModule { }
+export class AdminModule {
+}
