+++ frontend/src/app/admin/admin.routes.ts
@@ -2,7 +2,6 @@ import {RouterModule} from '@angular/router';
 import {OverviewComponent} from './overview/overview.component';
 import {OrdersComponent} from './orders/orders.component';
 import {PlansComponent} from './plans/plans.component';
-import {SidebarPageGroup} from '../ui/sidebar/sidebar.component';
 import {UsersComponent} from './users/users.component';
 
 export const adminRoutes = RouterModule.forChild([
@@ -11,43 +10,3 @@ export const adminRoutes = RouterModule.forChild([
   {path: 'users', component: UsersComponent},
   {path: 'plans', component: PlansComponent},
 ]);
-
-export const adminPages: SidebarPageGroup[] = [
-  {
-    title: 'Systemverwaltung',
-    pages: [
-      {
-        title: 'Statistiken',
-        icon: 'bar-chart-2',
-        url: '/admin/',
-        matchFull: true
-      },
-      {
-        title: 'Benutzer und Kunden',
-        icon: 'users',
-        url: '/admin/users',
-      },
-      {
-        title: 'Tarife',
-        icon: 'map',
-        url: '/admin/plans',
-      },
-      {
-        title: 'Bestellungen',
-        icon: 'shopping-cart',
-        url: '/admin/orders',
-      }
-    ]
-  },
-  {
-    title: 'Weitere Funktionen',
-    pages: [
-      {
-        title: 'Profil bearbeiten',
-        icon: 'user',
-        url: '/profile/',
-        matchFull: true
-      },
-    ]
-  }
-];
