+++ frontend/src/app/admin/orders/orders.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import {UiButtonGroup} from '../../ui/ui.interface';
 import {ApiService} from '../../services/api/api.service';
 import {OrderData} from '../../interfaces/order.interface';
+import {adminPages} from '../admin.pages';
 
 @Component({
   selector: 'app-orders',
@@ -9,6 +10,7 @@ import {OrderData} from '../../interfaces/order.interface';
   styleUrls: ['./orders.component.scss']
 })
 export class OrdersComponent implements OnInit {
+  public sidebarPages = adminPages;
   public results: OrderData[] = [];
   public loading = false;
   public buttonGroup: UiButtonGroup = {
