+++ frontend/src/app/admin/overview/overview.component.html
@@ -1,3 +1,3 @@
-<app-admin-layout>
+<app-layout>
   Test
-</app-admin-layout>
+</app-layout>
