+++ frontend/src/app/admin/plans/plans.component.html
@@ -1,4 +1,4 @@
-<app-admin-layout [buttons]="buttonGroup" title="Tarife bearbeiten">
+<app-layout [sidebarPages]="sidebarPages" [buttons]="buttonGroup" title="Tarife bearbeiten">
   <div class="bg-white rounded border-1 border-black p-4">
     <app-loading *ngIf="loading"></app-loading>
     <table class="table-fixed w-full" *ngIf="!loading">
@@ -27,7 +27,7 @@
       </tbody>
     </table>
   </div>
-</app-admin-layout>
+</app-layout>
 <ng-template #editModal>
   <label class="block mt-6 leading-3 text-lg">Identifier</label>
   <input class="mt-4 bg-gray-300 form-input block w-full rounded-md sm:leading-5" type="text" [(ngModel)]="currentEditPlan.id"
