+++ frontend/src/app/admin/users/users.component.html
@@ -1,4 +1,4 @@
-<app-admin-layout [buttons]="buttonGroup" title="Benutzer verwalten">
+<app-layout [sidebarPages]="sidebarPages" [buttons]="buttonGroup" title="Benutzer verwalten">
   <div class="bg-white rounded border-1 border-black p-4">
     <app-loading *ngIf="loading"></app-loading>
     <table class="table-fixed w-full" *ngIf="!loading">
@@ -27,4 +27,4 @@
       </tbody>
     </table>
   </div>
-</app-admin-layout>
+</app-layout>
