+++ frontend/src/app/ui/layout/layout.component.html
@@ -9,7 +9,7 @@
 <app-container [large]="true">
   <div class="md:flex">
     <div class="mb-3 md:mr-2" style="min-width: 300px">
-      <app-sidebar [pageGroups]="p"></app-sidebar>
+      <app-sidebar [pageGroups]="sidebarPages"></app-sidebar>
     </div>
     <div class="flex-1 md:ml-2">
       <ng-content></ng-content>
